TimeInput and TimeInputWithButton
