--- apps/client/src/common/components/input/time-input/TimeInputSimple.module.scss
@@ -1,11 +1,0 @@
-@use "../../../../theme/v2Styles" as *;
-
-$input-font-size: 15px;
-$input-delayed-border-color: #E69056;
-
-.timeInput {
-  font-size: $input-font-size;
-  letter-spacing: 1px;
-  width: 8em;
-}
-
