+++ apps/client/src/features/event-editor/composite/EventEditorTimes.tsx
@@ -4,7 +4,7 @@ import { EndAction, OntimeEvent, TimerType } from 'ontime-types';
 import { calculateDuration, millisToString } from 'ontime-utils';
 
 import TimeInput from '../../../common/components/input/time-input/TimeInput';
-import TimeInputSimple from '../../../common/components/input/time-input/TimeInputSimple';
+import TimeInputWithButton from '../../../common/components/input/time-input/TimeInputWithButton';
 import { useEventAction } from '../../../common/hooks/useEventAction';
 import { millisToDelayString } from '../../../common/utils/dateConfig';
 import { cx } from '../../../common/utils/styleUtils';
@@ -95,7 +95,7 @@ const EventEditorTimes = (props: EventEditorTimesProps) => {
         <label className={inputTimeLabels} htmlFor='timeStart'>
           {startLabel}
         </label>
-        <TimeInput
+        <TimeInputWithButton
           id='timeStart'
           name='timeStart'
           submitHandler={handleSubmit}
@@ -108,7 +108,7 @@ const EventEditorTimes = (props: EventEditorTimesProps) => {
         <label className={inputTimeLabels} htmlFor='timeEnd'>
           {endLabel}
         </label>
-        <TimeInput
+        <TimeInputWithButton
           id='timeEnd'
           name='timeEnd'
           submitHandler={handleSubmit}
@@ -121,7 +121,7 @@ const EventEditorTimes = (props: EventEditorTimesProps) => {
         <label className={style.inputLabel} htmlFor='durationOverride'>
           Duration
         </label>
-        <TimeInput
+        <TimeInputWithButton
           id='durationOverride'
           name='durationOverride'
           submitHandler={handleSubmit}
@@ -166,7 +166,7 @@ const EventEditorTimes = (props: EventEditorTimesProps) => {
         <label className={style.inputLabel} htmlFor='timeWarning'>
           Warning Time
         </label>
-        <TimeInputSimple
+        <TimeInput
           id='timeWarning'
           name='timeWarning'
           submitHandler={handleSubmit}
@@ -178,7 +178,7 @@ const EventEditorTimes = (props: EventEditorTimesProps) => {
         <label className={style.inputLabel} htmlFor='timeDanger'>
           Danger Time
         </label>
-        <TimeInputSimple
+        <TimeInput
           id='timeDanger'
           name='timeDanger'
           submitHandler={handleSubmit}
